Third Commit

+++ database/migrations/2022_03_17_144351_create_goal_forms_table.php
@@ -23,7 +23,7 @@ class CreateGoalFormsTable extends Migration
             $table->tinyInteger('quantity');
             $table->float('unitary_value');
             $table->float('amount');
-            $table->tinyInteger('year');
+            $table->integer('year');
             $table->timestamps();
 
             $table->engine = 'InnoDB';
